+++ packages/react-self-serializers/src/GraphQLComponent.tsx
@@ -1,30 +1,38 @@
 import React from 'react';
 
 import {useSerialized} from '@shopify/react-html';
-import {
-  ApolloProvider,
-  createGraphQLClient,
-  Options,
-} from '@shopify/react-graphql';
+import {ApolloProvider, createSsrExtractableLink} from '@shopify/react-graphql';
 import {useLazyRef} from '@shopify/react-hooks';
 
-interface Props extends Partial<Options> {
+interface Props {
   children?: React.ReactNode;
+  createClient(): import('apollo-client').ApolloClient<any>;
 }
 
-export function GraphQL({children, ...explicitOptions}: Props) {
-  const [initialData, Serialize] = useSerialized<Options['initialData']>(
-    'apollo',
-  );
+export function GraphQL({children, createClient}: Props) {
+  const [initialData, Serialize] = useSerialized<object | undefined>('apollo');
 
-  const client = useLazyRef(() =>
-    createGraphQLClient({initialData, ...explicitOptions}),
-  ).current;
+  const [client, link] = useLazyRef<
+    [
+      import('apollo-client').ApolloClient<any>,
+      ReturnType<typeof createSsrExtractableLink>
+    ]
+  >(() => {
+    const link = createSsrExtractableLink();
+    const client = createClient();
+    client.link = link.concat(client.link);
+
+    if (initialData) {
+      client.cache = client.cache.restore(initialData);
+    }
+
+    return [client, link];
+  }).current;
 
   return (
     <>
       <ApolloProvider client={client}>{children}</ApolloProvider>
-      <Serialize data={() => client.resolve()} />
+      <Serialize data={() => link.resolveAll(() => client.extract())} />
     </>
   );
 }